test: merge run coverage + consolidate UI + distribute small scattered files
Phase C: Merge and consolidate
- test_run_ingestion_parsers_coverage → test_run (ingestion/parser tests)
- test_ui_facade_coverage + test_ui_ux + test_ui_wrapper → new test_ui

Phase D: Distribute and delete small files
- test_simple_coverage → distributed to test_health, test_core
- test_schemas_coverage → test_provider_models
- test_supplementary_coverage → distributed to test_check, test_mcp_server,
  test_cli_unit, test_core, test_indexing_service, test_search_index
- test_example_using_helpers deleted (demo/tutorial, not production tests)

Net: -8 files, -10 tutorial-only tests. All production tests preserved.
